requirements.txt
django==1.8.1
ipdb==0.8
uwsgi==2.0.10
web.py
beautifulsoup4==4.3.2
requests==2.7.0
youtube_dl==2015.05.20
celery==3.1.17
django-celery==3.1.16
subprocess32==3.2.6
redis==2.10.3
mock==1.0.1
django-nose==1.4
django-cassandra-engine==0.3.3
coverage==4.0a5